+++ packages/spectacular/src/index.ts
@@ -8,7 +8,7 @@ export * from './lib/feature-test-harness/create-feature-test-harness-options';
 export * from './lib/feature-test-harness/feature-test-harness';
 
 // Pipe testing API
-export * from './lib/pipe-test-harness/pipe-host/spectacular-pipe-host.component';
-export * from './lib/pipe-test-harness/spectacular-create-pipe-options';
-export * from './lib/pipe-test-harness/spectacular-pipe-fixture';
-export * from './lib/pipe-test-harness/spectacular-pipe-testbed';
+export * from './lib/pipe-testing/pipe-host/spectacular-pipe-host.component';
+export * from './lib/pipe-testing/spectacular-create-pipe-options';
+export * from './lib/pipe-testing/spectacular-pipe-fixture';
+export * from './lib/pipe-testing/spectacular-pipe-testbed';
