+++ packages/spectacular/src/lib/pipe-testing/common-pipes.integration.spec.ts
