+++ packages/spectacular/src/lib/pipe-testing/pipe-host/create-pipe-host-template.ts
