+++ packages/spectacular/src/lib/pipe-testing/pipe-host/internal-pipe-host-text-id.ts
