+++ packages/spectacular/src/lib/pipe-testing/pipe-host/spectacular-pipe-host.component.ts
