+++ packages/spectacular/src/lib/pipe-testing/spectacular-create-pipe-options.ts
