+++ packages/spectacular/src/lib/pipe-testing/test-util/create-common-pipe-options.ts
